separate comboboxdialog
